requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
PyMySQL>=1.0.0
python-decouple>=3.6
PyJWT>=2.6.0
cryptography>=40.0.0
django-extensions>=3.2.0
Pillow>=9.0.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
requests>=2.28.0